+++ models/dao/OBAModelDAO.m
@@ -126,7 +126,7 @@ const static int kMaxEntriesInMostRecentList = 10;
     existingEvent.title = event.title;
     existingEvent.subtitle = event.subtitle;
     
-    int over = [_mostRecentStops count] - kMaxEntriesInMostRecentList;
+    NSInteger over = [_mostRecentStops count] - kMaxEntriesInMostRecentList;
     for( int i=0; i<over; i++)
         [_mostRecentStops removeObjectAtIndex:([_mostRecentStops count]-1)];
     
@@ -154,7 +154,7 @@ const static int kMaxEntriesInMostRecentList = 10;
         
     }
     
-    int over = [_mostRecentCustomApiUrls count] - kMaxEntriesInMostRecentList;
+    NSUInteger over = [_mostRecentCustomApiUrls count] - kMaxEntriesInMostRecentList;
     for( int i=0; i<over; i++)
         [_mostRecentCustomApiUrls removeObjectAtIndex:([_mostRecentCustomApiUrls count]-1)];
